Add basic functionality for discussions API endpoints
