+++ library/Garden/Web/Exception/HttpException.php
@@ -81,7 +81,7 @@ abstract class HttpException extends \Exception implements \JsonSerializable {
      * - TODO: Keys beginning with **HTTP_** will be added as headers.
      * - **description** will give the exception a longer description.
      *
-     * @param string $message The error message.
+     * @param int|string $message The error message.
      * @param int $code The http error code.
      * @param array $context An array of context variables that can be used to render a more detailed response.
      */
